+++ src/compiler/file_uploader/uploader_actor.ts
@@ -1,5 +1,5 @@
 import { Actor, ActorMethod, ActorSubclass } from '@dfinity/agent';
-import { createAuthenticatedAgent } from '../../../../dfx';
+import { createAuthenticatedAgent } from '../../../dfx';
 
 export async function createActor(
     canisterId: string
